+++ ESFramework/Foundation/NSArray+ESExtension.h
@@ -28,6 +28,8 @@ NS_ASSUME_NONNULL_BEGIN
 - (nullable ObjectType)nextObjectToIndex:(NSUInteger)index;
 - (nullable ObjectType)nextObjectToObject:(ObjectType)object;
 
+- (nullable NSData *)JSONData;
+- (nullable NSData *)JSONDataWithOptions:(NSJSONWritingOptions)opts;
 - (nullable NSString *)JSONString;
 - (nullable NSString *)JSONStringWithOptions:(NSJSONWritingOptions)opts;