+++ ESFramework/Foundation/NSDictionary+ESExtension.m
@@ -43,6 +43,18 @@
     return [NSDictionary dictionaryWithObjects:objects forKeys:keys];
 }
 
+- (NSData *)JSONData
+{
+    return [self JSONDataWithOptions:0];
+}
+
+- (NSData *)JSONDataWithOptions:(NSJSONWritingOptions)opts
+{
+    return [NSJSONSerialization isValidJSONObject:self]
+    ? [NSJSONSerialization dataWithJSONObject:self options:opts error:NULL]
+    : nil;
+}
+
 - (NSString *)JSONString
 {
     return [self JSONStringWithOptions:0];
@@ -50,9 +62,7 @@
 
 - (NSString *)JSONStringWithOptions:(NSJSONWritingOptions)opts
 {
-    return [NSJSONSerialization isValidJSONObject:self]
-           ? [NSJSONSerialization dataWithJSONObject:self options:opts error:NULL].UTF8String
-           : nil;
+    return [self JSONDataWithOptions:opts].UTF8String;
 }
 
 @end